Squiz / FunctionComment: Fix fixer for param with multiple variable types.
When the fixer encountered incorrect param types and there were multiple types, like `array()|int`, the param type would be replaced with the correction for the first incorrect param type on the line. Any of the other param types would be lost.

Example:

Test case:
```php
 * @param MyClass|int  $a4 Comment here.
```

Was incorrectly being fixed to read:
```php
 * @param integer $a4 Comment here.
```

And will now fix correctly to:
```php
 * @param MyClass|integer $a4 Comment here.
```