Fixer: stricter success determination
When the fixer reaches the 50 loops, it basically always means there is a fixer conflict.
The last loop *may* or *may not* have successfully fixed some issues, resulting in "no fixes" left, however, this can be an interim state.

For example, try running the `Squiz` standard over the `UnitTest.inc` file. It will report that the file was fixed successfully, however, when run with `-vv` it shows that the there was actually a fixer conflict and that the file therefore can not be considered 100% fixed.

This PR changes whether the Fixer reports success/failure by also taking the number of loops used into consideration and causes any file where the Fixer needed 50 loops to be reported as "Failed to fix".