Refactor/NameUtils: improve parse error/live coding resilience for the getDeclarationName() method
The `getDeclarationName()` method could inadvertently misidentify the name of a structure in the case of parse errors/live coding as it looked for the first `T_STRING` after the keyword without limiting it to this statement or anything, not just at the first non empty token.

This has now been fixed.

Included additional unit tests guarding against this.