Generators/Markdown: fix line break and indentation handling
As things were, line breaks in a `<standard>` block were not respected for proper display in markdown.

As a side-effect of this, an indented text line in a `<standard>` block directly after a blank line, would unintentionally cause the markdown to display the line as a code block (via the indentation).

This has now been fixed by:
* Trimming indentation off each line.
* Adding two spaces to the end of each line as long as the line is not followed by a blank line or at the end of the standard block.
    The two spaces will ensure markdown will recognize the line break (in most markdown flavours).

Includes updated test expectations.

Refs:
* https://www.markdownguide.org/basic-syntax/#paragraph-best-practices
* https://www.markdownguide.org/basic-syntax/#line-break-best-practices
* https://daringfireball.net/projects/markdown/syntax#p