Generators HTML/Markdown: use sentence case for doc titles
As things were, the "title" for a document would be printed as-is in Markdown/HTML.

For improved readability, this is now changed to "sentence case", i.e. always capitalize the first character of the title.

Note: This only changes the display title, the title used for creating the link used in the HTML Table of Contents is not changed.

Includes updated test expectations.