PSR2/NamespaceDeclaration: do not enforce new line in inline HTML
If a namespace declaration has a PHP close tag on the same line, the new line would be added as `T_INLINE_HTML` and therefore not be recognized as a blank line as the sniff solely looks for `T_WHITESPACE`. This then leads to a fixer conflict where the sniff just keeps adding new lines until it runs out of loops.

```
        => Fixing file: 0/6 violations remaining [made 46 passes]...
        * fixed 0 violations, starting loop 47 *
        PSR2.Namespaces.NamespaceDeclaration:81 replaced token 118 (T_INLINE_HTML on line 32) "\n\n" => "\n\n\n"
        PSR2.Namespaces.NamespaceDeclaration:81 replaced token 153 (T_INLINE_HTML on line 58) "\n\n" => "\n\n\n"
        => Fixing file: 2/6 violations remaining [made 47 passes]...
        * fixed 2 violations, starting loop 48 *
        **** PSR2.Namespaces.NamespaceDeclaration:81 has possible conflict with another sniff on loop 46; caused by the following change ****
        **** replaced token 118 (T_INLINE_HTML on line 32) "\n\n" => "\n\n\n" ****
        **** ignoring all changes until next loop ****
        => Fixing file: 0/6 violations remaining [made 48 passes]...
        * fixed 0 violations, starting loop 49 *
```

In my opinion, the sniff should bow out in that situation and should not enforce a new line as it may impact HTML display.

This commit implements this.

Includes a test safeguarding the fix.