Squiz/FunctionComment: selectively fix parse error in test case file
While there are definitely a lot more of these type of parse errors (method modifiers applied to global function) in this test case file, only one is problematic in the context of an automated fixer conflict check.

This commit fixes that one instance only.

As the others are not problematic, they are left as they are to prevent accidentally removing valid test situations from the test case file.