Squiz/FunctionSpacing: bug fix for attribute handling
The `Squiz.WhiteSpace.FunctionSpacing` sniff is supposed to check the number of blank lines before (and after) a function declaration, while taking docblocks and attributes attached to the function into account. I.e. if there is a docblock and/or attributes, the blank lines _above_ those will be checked.

Additionally, the sniff has a separate error code `BeforeFirst` for the number of blank lines before a first function in a class and the expected number of blank lines for the first function in a class may also be configured to be different than the "normal" number of blank lines before a function/between methods.

Now, while the sniff does take attributes into account, it does so incorrectly:
1. It doesn't allow for attributes _before_ a docblock, even though this is perfectly valid PHP and the attributes still belong with the function.
2. It doesn't allow for multiple attributes on the same line. Again, this is perfectly valid PHP.

The effect of this bug can be seen by checking the new tests without the fix:
1. The "Correct" functions will also be flagged as incorrect as the sniff gets confused by multiple attributes on the same line.
2. The set of tests with the attributes before the docblock will be flagged with the `Before` error code instead of `BeforeFirst` as the functions will not be recognized as the first in class due to the attributes before the docblock.
3. The fixer will make incorrect fixes.

Fixed now by removing the presumption about the _order_ of docblocks vs attributes and stabilizing the "current line" determination.

Includes tests.