Squiz/MemberVarSpacing: fix some unintentional errors in the test case file
These changes do not diminish the value of the tests, but prevent a parse error and/or fatal if the test case file would be linted.

Note: this doesn't mean the file lints cleanly now, but it's better than it was (there are still some interfaces with properties in the file).