Squiz/MemberVarSpacing: bug fix for attribute handling
The `Squiz.WhiteSpace.MemberVarSpacing` sniff is supposed to check the number of blank lines before a property declaration, while taking docblocks and attributes attached to the property into account. I.e. if there is a docblock and/or attributes, the blank lines _above_ those will be checked.

Additionally, the sniff has a separate error code `FirstIncorrect` for the number of blank lines before the first property in a class and the expected number of blank lines for the first property in a class may also be configured to be different than the "normal" number of blank lines before a property.

Now, while the sniff does take attributes into account, it does so incorrectly, as it doesn't allow for attributes _before_ a docblock, even though this is perfectly valid PHP and the attributes still belong with the property.

The effect of this bug can be seen by checking the new tests without the fix:
1. The "Correct" property will be flagged as incorrect as the sniff does determine the start of the property "preamble" correctly.
2. For the number of blank lines before, these will all be flagged with the `Incorrect` error code instead of `FirstIncorrect` as the properties will not be recognized as the first in class due to the attributes before the docblock.
3. The fixer will make incorrect fixes.

Fixed now by removing the presumption about the _order_ of docblocks vs attributes and stabilizing the "start of preamble" determination.

Includes tests.

Note: now that attributes before a docblock a also taken into account (also for the "blank line between bits of the preamble" determination), the `AfterComment` error code isn't correct anymore.
Changing this code, however, is a breaking change and as the message for the error has been updated already, not terrible urgent, so I propose to change error code to `BlankLineInPreamble` in PHPCS 4.0.