4.0 | Squiz/MemberVarSpacing: change error code
Since PR ..., the check for blank lines in the preamble, doesn't just check for blank lines between a docblock and the declaration, but also for blank lines between attributes and docblocks if the attribute would come before the docblock.

This meant that the error code is no longer covering the reality of what the check does.

This commit changes the error code from `AfterComment` to `BlankLineInPreamble` to fix that.